fix: move checking first_event_at_empty flow to stats api
